refactor(deploy): drop ./api.ts and ./mock.ts test indirection
Switch deploy command imports back to lib/plapi.ts directly and remove the
./mock.ts harness plus the --test-force-* / --test-fail-* CLI flags it
backed. The wrappers were added for an earlier mockable-deploy-API
experiment that is no longer needed.
